1. 将demo data与common.go移到model目录中 2. 将common.go中的Video改名为VideoFlow，与数据库中的Video区分 3. 推流接口基本实现（需要与User,favorite对接）

+++ model/video.go
@@ -8,8 +8,8 @@ type Video struct {
 	Title         string
 	PlayUrl       string
 	CoverUrl      string
-	FavoriteCount int
-	CommentCount  int
+	FavoriteCount int64
+	CommentCount  int64
 	Status        byte
 	HashValue     string `gorm:"column:hash"`
 	CreateTime    time.Time
